slackBot/requirements.txt
slack-bolt>=1.18.0
slack-sdk>=3.26.0
python-dotenv>=1.0.0
requests>=2.31.0
